backend/requirements.txt
Django==3.1.5
gunicorn==20.0.4
psycopg2-binary==2.8.6
djangorestframework==3.12.2
django-filter==2.4.0
django-url-filter==0.3.15
django-settings-export==1.2.1
autopep8==1.5.4
pycodestyle==2.6.0
pylint==2.6.0
python-dateutil==2.8.1
